UInput: Move implementation to separate file, expose C++ interface.
The implementation is moved to a separate file to prevent pulling in
common.h when needing just the uinput functionality.
